requirements.txt
torch
torchvision
fastapi
uvicorn
pillow
scikit-image